// src/modules/content/model/content.model.ts
export class ContentModel {
  id: number;
  project_id: string;
  post_type: string;
  advice?: string;
  benefit?: string;
  platform: string;
  tone: string;
  length: string;
  use_emojis: boolean;
  use_hashtags: boolean;
  content?: string;
  status: string;
  created_at: string;
  updated_at?: string;
}